+++ ChangeLog.md
@@ -20,6 +20,7 @@
 - Added a "limit" parameter to the last events controller (dashboard/events)
 - Records that are not published will be displayed with a yellow background on the record list
 - Added a third parameter (per_page) to the record_list function
+- Added a "note" attribute to the description node of each field
 
 
 **v 0.8.4** (2011-10-25)